+++ backend/backend/celery.py
@@ -62,12 +62,12 @@ def receiver_setup_logging(loglevel, logfile, format, colorize, **kwargs):  # pr
 
 @app.on_after_configure.connect
 def setup_periodic_tasks(sender, **kwargs):
+    from substrapp.tasks.tasks_compute_task import prepare_aggregate_task
+    from substrapp.tasks.tasks_compute_task import prepare_composite_training_task
+    from substrapp.tasks.tasks_compute_task import prepare_testing_task
+    from substrapp.tasks.tasks_compute_task import prepare_training_task
     from substrapp.tasks.tasks_docker_registry import clean_old_images_task
     from substrapp.tasks.tasks_docker_registry import docker_registry_garbage_collector_task
-    from substrapp.tasks.tasks_prepare_task import prepare_aggregate_task
-    from substrapp.tasks.tasks_prepare_task import prepare_composite_training_task
-    from substrapp.tasks.tasks_prepare_task import prepare_testing_task
-    from substrapp.tasks.tasks_prepare_task import prepare_training_task
 
     period = int(settings.CELERYBEAT_SCHEDULE_TASK_PERIOD)