+++ backend/substrapp/tasks/__init__.py
@@ -2,5 +2,4 @@
 from .tasks_compute_plan import *
 from .tasks_compute_task import *
 from .tasks_docker_registry import *
-from .tasks_prepare_task import *
 from .tasks_remove_intermediary_models import *